- A new field added to the "settings.json" files indicating what format are the settings based on (currently supporting Lua). - Divided the "main.py" representing the main GUI into two separate files second of which implements the second window. - Some functionalities improved (like closing the second window using canceel button and restructured some of the functions).
