+++ src/dependency/Pyegi/PythonScripts/Disintegration/settings.json
@@ -1,4 +1,5 @@
 {
+  "Format": "Lua",
   "Controls": [
     {
       "x": 0,
